+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/library/LibraryData.java
@@ -48,6 +48,7 @@ import javax.lang.model.element.TypeElement;
 import javax.lang.model.type.TypeMirror;
 
 import com.oracle.truffle.dsl.processor.ProcessorContext;
+import com.oracle.truffle.dsl.processor.java.ElementUtils;
 import com.oracle.truffle.dsl.processor.model.MessageContainer;
 import com.oracle.truffle.dsl.processor.model.Template;
 
@@ -64,9 +65,18 @@ public final class LibraryData extends Template {
     private TypeMirror assertions;
 
     private ExportsLibrary objectExports;
+    private boolean defaultExportLookupEnabled;
 
     public LibraryData(TypeElement type, AnnotationMirror annotationMirror) {
         super(ProcessorContext.getInstance(), type, annotationMirror);
+    }
+
+    public void setDefaultExportLookupEnabled(boolean defaultExportLookupEnabled) {
+        this.defaultExportLookupEnabled = defaultExportLookupEnabled;
+    }
+
+    public boolean isDefaultExportLookupEnabled() {
+        return defaultExportLookupEnabled;
     }
 
     public void setExportsReceiverType(TypeMirror receiverType) {
@@ -137,6 +147,21 @@ public final class LibraryData extends Template {
 
     public ExportsLibrary getObjectExports() {
         return objectExports;
+    }
+
+    public LibraryDefaultExportData getBuiltinDefaultExport(TypeMirror receiverType) {
+        for (LibraryDefaultExportData export : defaultExports) {
+            if (export.isDefaultObjectExport()) {
+                continue;
+            }
+            if (ElementUtils.isAssignable(export.getReceiverType(), receiverType)) {
+                return export;
+            }
+            if (ElementUtils.isAssignable(ProcessorContext.getInstance().getType(Object.class), export.getReceiverType())) {
+                return export;
+            }
+        }
+        return null;
     }
 
 }