+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/library/LibraryParser.java
@@ -94,6 +94,8 @@ public class LibraryParser extends AbstractParser<LibraryData> {
             return model;
         }
 
+        model.setDefaultExportLookupEnabled(ElementUtils.getAnnotationValue(Boolean.class, mirror, "defaultExportLookupEnabled"));
+
         boolean defaultExportReachable = true;
         List<AnnotationMirror> defaultExports = ElementUtils.getRepeatedAnnotation(element.getAnnotationMirrors(), types.GenerateLibrary_DefaultExport);
         for (AnnotationMirror defaultExport : defaultExports) {
@@ -119,6 +121,7 @@ public class LibraryParser extends AbstractParser<LibraryData> {
                 defaultExportReachable = false;
             }
         }
+
         parseAssertions(element, mirror, type, model);
 
         List<ExecutableElement> allMethods = ElementFilter.methodsIn(CompilerFactory.getCompiler(type).getEnclosedElementsInDeclarationOrder(type));